+++ apps/preview/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ensembleui/react-preview
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [357d754]
+  - @ensembleui/react-framework@0.1.32
+  - @ensembleui/react-runtime@0.1.42
+
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [3da9f12]
+- Updated dependencies [f566152]
+- Updated dependencies [f566152]
+  - @ensembleui/react-runtime@0.1.41
+  - @ensembleui/react-framework@0.1.31
+
 ## 0.0.29
 
 ### Patch Changes