+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ensembleui/react-framework
 
+## 0.1.32
+
+### Patch Changes
+
+- 357d754: Fix context passing on imperative invoke API
+
+## 0.1.31
+
+### Patch Changes
+
+- f566152: Optionally debounce state updates for rapid state changes
+- f566152: Support env variables in API uris
+
 ## 0.1.30
 
 ### Patch Changes